made PR changes
