+++ pilco/policies/eq_policy.py
@@ -29,19 +29,19 @@ class EQPolicy(Policy):
         # Number of radial basis functions
         self.num_eq_features = num_eq_features
 
-        # Set RBF policy locations
+        # Set EQ policy locations
         eq_locs_init = tf.zeros((num_eq_features, state_dim), dtype=dtype)
         self.eq_locs = UnconstrainedVariable(eq_locs_init,
                                              name='eq_locs',
                                              dtype=dtype)
 
-        # Set RBF policy lengthscales
+        # Set EQ policy lengthscales
         eq_log_scales_init = tf.zeros((1, state_dim), dtype=dtype)
         self.eq_log_scales = UnconstrainedVariable(eq_log_scales_init,
                                                     name='eq_log_scales',
                                                     dtype=dtype)
 
-        # Set RBF policy weights
+        # Set EQ policy weights
         eq_weights_init = tf.zeros((num_eq_features,), dtype=dtype)
         self.eq_weights = UnconstrainedVariable(eq_weights_init,
                                                 name='eq_weights',
@@ -162,7 +162,7 @@ class EQPolicy(Policy):
     @property
     def eq_scales(self):
         """
-        Returns 1 x state_dim tensor of RBF squared lengthscales.
+        Returns 1 x state_dim tensor of EQ squared lengthscales.
         """
 
         return tf.math.exp(self.eq_log_scales()) ** 2
@@ -172,7 +172,7 @@ class EQPolicy(Policy):
     def eq_scales_mat(self):
         """
         Returns diagonal state_dim x state_dim tensor
-        of RBF squared lengthscales.
+        of EQ squared lengthscales.
         """
 
         eq_scales = self.eq_scales[0]
@@ -195,62 +195,9 @@ class EQPolicy(Policy):
 
         exp_quad = tf.math.exp(-0.5 * quad)
 
-        # RBF output is the weighted sum of eq components
+        # EQ output is the weighted sum of eq components
         eq = tf.einsum('i, i ->',
                         self.eq_weights(),
                         exp_quad)
 
         return eq
-
-
-
-class PendulumEQPolicy(EQPolicy):
-
-
-    def __init__(self,
-                 num_eq_features,
-                 dtype,
-                 name='pendulum_eq_policy',
-                 **kwargs):
-
-        super().__init__(state_dim=2,
-                         action_dim=1,
-                         num_eq_features=num_eq_features,
-                         dtype=dtype,
-                         **kwargs)
-
-
-    def reset(self):
-
-        """
-        Resets policy parameters: location, scale and weights of EQ features.
-        :return:
-        """
-
-        # Locations of EQ scales along the angle dimension
-        eq_theta_locs = tf.random.normal(shape=(self.num_eq_features,),
-                                         mean=-np.pi,
-                                         stddev=2 * np.pi,
-                                         dtype=self.dtype)
-
-        # Locations of EQ scales along the angular velocity dimension
-        eq_theta_dot_locs = tf.random.normal(shape=(self.num_eq_features,),
-                                             mean=0.,
-                                             stddev=4.,
-                                             dtype=self.dtype)
-
-        # Concatenate components of EQ locations and assign
-        eq_locs = tf.stack([eq_theta_locs,
-                            eq_theta_dot_locs], axis=-1)
-        self.eq_locs.assign(eq_locs)
-
-        # Log-scales of EQ components (shared across all EQ components)
-        eq_log_scales = self.num_eq_features ** -0.5 * tf.ones(shape=(1, self.state_dim), dtype=self.dtype)
-        self.eq_log_scales.assign(eq_log_scales)
-
-        # Weight coefficients for each EQ component
-        eq_weights_init = tf.random.normal(shape=(self.num_eq_features,),
-                                           mean=0.,
-                                           stddev=0.3,
-                                           dtype=self.dtype)
-        self.eq_weights.assign(eq_weights_init)
